+++ osu.Game.Rulesets.Taiko/Objects/Drawables/DrawableStrongHandler.cs
@@ -2,16 +2,20 @@
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
 using osu.Game.Rulesets.Objects.Drawables;
+using osu.Game.Rulesets.Taiko.Judgements;
 
 namespace osu.Game.Rulesets.Taiko.Objects.Drawables
 {
-    public abstract class DrawableStrongHitObject : DrawableTaikoHitObject
+    /// <summary>
+    /// Used as a nested hitobject to provide <see cref="TaikoStrongJudgement"/>s for <see cref="DrawableTaikoHitObject"/>s.
+    /// </summary>
+    public abstract class DrawableStrongHandler : DrawableTaikoHitObject
     {
         public override bool DisplayJudgement => false;
 
         public readonly DrawableHitObject MainObject;
 
-        protected DrawableStrongHitObject(StrongHitObject strong, DrawableHitObject mainObject)
+        protected DrawableStrongHandler(StrongHitObject strong, DrawableHitObject mainObject)
             : base(strong)
         {
             MainObject = mainObject;
